Doc2Command/requirements.txt
Pillow==9.2.0
torch==2.0.1
tqdm
transformers==4.33.1
einops==0.7.0